fix: applying PR feedback
