+++ frontend/lib/components/drawer.dart
@@ -5,10 +5,10 @@ import 'package:frontend/components/router.dart';
 import 'package:frontend/main.dart';
 import 'package:frontend/models/member.dart';
 import 'package:frontend/routes/UnknownScreen.dart';
-import 'package:frontend/routes/company/ItemPackagePage.dart';
+import 'package:frontend/routes/items_packages/ItemPackagePage.dart';
 import 'package:frontend/routes/meeting/MeetingPage.dart';
-import 'package:frontend/routes/member/MemberPage.dart';
-import 'package:frontend/routes/member/MemberScreen.dart';
+import 'package:frontend/routes/members_teams/member/MemberPage.dart';
+import 'package:frontend/routes/members_teams/member/MemberScreen.dart';
 import 'package:frontend/routes/session/SessionPage.dart';
 import 'package:frontend/services/authService.dart';
 import 'package:provider/provider.dart';
