+++ frontend/lib/routes/company/billing/AddBillingForm.dart
@@ -27,8 +27,9 @@ class _AddBillingFormState extends State<AddBillingForm> {
   final _eventController = TextEditingController();
   final _invoiceNumberController = TextEditingController();
   final _notesController = TextEditingController();
-  final _valueEurosController = TextEditingController();
-  final _valueCentsController = TextEditingController();
+  final _costController = TextEditingController();
+
+  late NumberFormat formatter;
 
   bool invoice = false;
   bool paid = false;
@@ -98,8 +99,10 @@ class _AddBillingFormState extends State<AddBillingForm> {
       var event = int.parse(_eventController.text);
       var invoiceNumber = _invoiceNumberController.text;
       var notes = _notesController.text;
-      var value = int.parse(_valueEurosController.text) * 100 +
-          int.parse(_valueCentsController.text);
+      var parseCost = double.parse(_costController.text);
+      var euros = parseCost.toInt();
+      var cents = ((parseCost - euros) * 100).round();
+      int cost = euros * 100 + cents;
 
       ScaffoldMessenger.of(context).showSnackBar(
         const SnackBar(
@@ -117,7 +120,7 @@ class _AddBillingFormState extends State<AddBillingForm> {
           paid: paid,
           proForma: proForma,
           receipt: receipt,
-          value: value,
+          value: cost,
           visible: visible);
 
       if (c != null) {
@@ -228,57 +231,25 @@ class _AddBillingFormState extends State<AddBillingForm> {
               ),
             ),
           ),
-          Row(
-            children: [
-              Expanded(
-                child: Padding(
-                  padding: const EdgeInsets.all(8.0),
-                  child: TextFormField(
-                    controller: _valueEurosController,
-                    validator: (value) {
-                      if (value == null || value.isEmpty) {
-                        return 'Please enter the cost of the billing';
-                      }
-                      int val = int.parse(value);
-                      if (val < 0 || val > 100) {
-                        return 'Number must be between 0 and 100';
-                      }
-                      return null;
-                    },
-                    decoration: const InputDecoration(
-                      icon: const Icon(Icons.money),
-                      labelText: "Cost of billing (only euros) *",
-                    ),
-                    keyboardType: TextInputType.number,
-                    inputFormatters: <TextInputFormatter>[
-                      FilteringTextInputFormatter.digitsOnly
-                    ],
-                  ),
-                ),
+          Padding(
+            padding: const EdgeInsets.all(8.0),
+            child: TextFormField(
+              controller: _costController,
+              validator: (value) {
+                if (value == null || value.isEmpty) {
+                  return 'Please enter the cost of the billing';
+                }
+                return null;
+              },
+              decoration: const InputDecoration(
+                icon: const Icon(Icons.money),
+                labelText: "Cost of billing *",
               ),
-              Expanded(
-                child: Padding(
-                  padding: const EdgeInsets.all(8.0),
-                  child: TextFormField(
-                    controller: _valueCentsController,
-                    validator: (value) {
-                      if (value == null || value.isEmpty) {
-                        return 'Please enter the cost of the billing';
-                      }
-                      return null;
-                    },
-                    decoration: const InputDecoration(
-                      icon: const Icon(Icons.money),
-                      labelText: "Cost of billing (only cents) *",
-                    ),
-                    keyboardType: TextInputType.number,
-                    inputFormatters: <TextInputFormatter>[
-                      FilteringTextInputFormatter.digitsOnly
-                    ],
-                  ),
-                ),
-              ),
-            ],
+              keyboardType: TextInputType.numberWithOptions(decimal: true),
+              inputFormatters: <TextInputFormatter>[
+                FilteringTextInputFormatter.allow(RegExp(r'^\d+\.?\d{0,2}'))
+              ],
+            ),
           ),
           CheckboxListTile(
             value: this.invoice,
