+++ frontend/lib/routes/members_teams/member/MemberPage.dart
@@ -4,9 +4,9 @@ import 'package:frontend/components/appbar.dart';
 import 'package:frontend/components/router.dart';
 import 'package:frontend/main.dart';
 import 'package:frontend/models/team.dart';
-import 'package:frontend/routes/member/MemberListWidget.dart';
-import 'package:frontend/routes/teams/TeamsNotifier.dart';
-import 'package:frontend/routes/teams/TeamsTable.dart';
+import 'package:frontend/routes/members_teams/member/MemberListWidget.dart';
+import 'package:frontend/routes/members_teams/teams/TeamsNotifier.dart';
+import 'package:frontend/routes/members_teams/teams/TeamsTable.dart';
 import 'package:frontend/services/authService.dart';
 import 'package:frontend/services/teamService.dart';
 import 'package:provider/provider.dart';
