+++ frontend/lib/routes/members_teams/member/MemberScreen.dart
@@ -4,9 +4,9 @@ import 'package:frontend/components/eventNotifier.dart';
 import 'package:frontend/components/memberPartCard.dart';
 import 'package:frontend/main.dart';
 import 'package:frontend/models/member.dart';
-import 'package:frontend/routes/member/DisplayContact2.dart';
-import 'package:frontend/routes/member/EditMemberForm.dart';
-import 'package:frontend/routes/member/MemberNotifier.dart';
+import 'package:frontend/routes/members_teams/member/DisplayContact2.dart';
+import 'package:frontend/routes/members_teams/member/EditMemberForm.dart';
+import 'package:frontend/routes/members_teams/member/MemberNotifier.dart';
 import 'package:frontend/services/authService.dart';
 import 'package:frontend/services/memberService.dart';
 import 'package:frontend/services/teamService.dart';
